server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
huggingface-hub==0.23.0
python-dotenv==1.0.1
Werkzeug==3.0.1
numpy>=1.23.5,<2.0.0
pillow>=10.3.0
tflite-runtime==2.14.0
setuptools
gunicorn>=21.2.0